requirements.txt
azure-functions
azure-cosmosdb-table